Improved Scheduling
I was just having a fucking brainfart there. Made an observer pattern core to the beat triggers, which is a much cleaner way to wire up. This way, when we change voices, we'll be able to just unsub/resub the voice to the right track. Also, voices can be multisubbed, which is good for letting the same voice appear on multiple tracks.

This is how I should have done this the first time.